Refactored `src/Tests/*`
