+++ src/Tests/V1/DatumInstances-test.ts
@@ -1,5 +1,5 @@
 // Alias the generator for `Datum` since `Datum` is a type alias for
 // `PlutusData`
-import * as TestPlutusData from "./PlutusDataInstances-test.js";
+import * as TestPlutusData from "../PlutusDataInstances-test.js";
 
 export const fcDatum = TestPlutusData.fcPlutusData;
