+++ src/Tests/V1/ScriptHashInstances-test.ts
@@ -1,10 +1,10 @@
 // Tests for the instances for `ScriptHash`
-import * as V1 from "../Lib/V1.js";
+import * as V1 from "../../Lib/V1.js";
 import * as Prelude from "prelude";
 
 import { describe, it } from "node:test";
 
-import * as TestUtils from "./TestUtils.js";
+import * as TestUtils from "../TestUtils.js";
 import fc from "fast-check";
 
 // `scriptHash1` is distinct from `scriptHash2`
